fix(web): make TaskList's SWR fetcher handle auth/server errors
Extract the auth/error-handling fetcher used by the dashboard page into
a shared src/lib/swr-fetcher.ts (missing token -> redirect to login,
401 -> clear auth + redirect, other non-ok -> throw so SWR exposes
error), and have TaskList use it and render a visible error state
instead of silently degrading to an empty task list.
